+++ src/dom/async_html.ts
@@ -1,13 +1,18 @@
-import { uid } from '../utils/uid';
+import { openHTMLTagRegExp } from '../utils/regexps';
 
 const DATA_ASYNC_HTML_STUB_ATTRIBUTE_NAME = 'data-async-html-stub';
 
 export const asyncHTMLMap: Map<string, Promise<string>> = new Map();
 
-export function createStub(asyncHTMLContainer: Promise<string>): string {
-  const stub = `__${uid()}__`;
-  asyncHTMLMap.set(stub, asyncHTMLContainer);
-  return `<div ${DATA_ASYNC_HTML_STUB_ATTRIBUTE_NAME}="${stub}"></div>`;
+export function createStub(stubId: string, stubComponent: string): string {
+  if (!stubComponent.trimStart().startsWith('<')) {
+    return `<div ${DATA_ASYNC_HTML_STUB_ATTRIBUTE_NAME}="${stubId}">${stubComponent}</div>`;
+  } else {
+    return stubComponent.replace(
+      openHTMLTagRegExp,
+      `$& ${DATA_ASYNC_HTML_STUB_ATTRIBUTE_NAME}="${stubId}"`
+    );
+  }
 }
 
 export function getStubElement(