+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './state/state';
+export * from './template/future';
 export * from './template/template';
 export * from './component/web_component';
 export * as i18n from './intl/i18n';