+++ src/utils/regexps.ts
@@ -5,3 +5,5 @@ export const booleanAttributeRegExp = /\?([\w-]+)=$/;
 export const hookAttributeRegExp = /:(mounted|updated|removed)=$/;
 
 export const specialPropertiesRegExp = /\.([\w-]+)=$/;
+
+export const openHTMLTagRegExp = /<[^\s>]+/;